@@ -2,27 +2,37 @@ apply plugin: 'com.android.application'
 
 android {
 
-    compileSdkVersion 23
-    buildToolsVersion "24.0.1"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.sola.v2ex_android"
-        minSdkVersion 19
-        targetSdkVersion 23
+        minSdkVersion 17
+        targetSdkVersion 28
         versionCode 2
         versionName "1.1"
         manifestPlaceholders = [CHANNEL_NAME: "MAIN"]
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         buildConfigField "boolean", "LOG_DEBUG", "false"
+        flavorDimensions "versionCode"
         vectorDrawables.useSupportLibrary = true
+    }
+
+    lintOptions {
+        checkReleaseBuilds false
+        abortOnError false
     }
 
     signingConfigs {
         config  {
-            keyAlias 'main'
-            keyPassword '15913878451'
-            storeFile file('D:/keystore/main.jks')
-            storePassword '15913878451'
+            keyAlias 'key0'
+            keyPassword '123456'
+            storeFile file('C:\\Work\\v2ex\\V2EX-Android\\app\\release.jks')
+            storePassword '123456'
         }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     buildTypes {
