@@ -1,10 +1,8 @@
 package com.sola.v2ex_android.ui.base;
 
 import android.os.Handler;
-import android.support.v4.widget.SwipeRefreshLayout;
-
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import com.sola.v2ex_android.R;
-
 import java.lang.ref.WeakReference;
 
 /**
