@@ -1,10 +1,11 @@
 package com.sola.v2ex_android.ui.base.adapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.sola.v2ex_android.ui.base.callback.OnItemClickListener;
 import com.sola.v2ex_android.util.ValidateUtil;
