@@ -5,11 +5,12 @@ package com.sola.v2ex_android.ui.base.adapter;
  */
 
 import android.content.Context;
-import android.support.annotation.LayoutRes;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.LayoutRes;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.rockerhieu.rvadapter.RecyclerViewAdapterWrapper;
 import com.sola.v2ex_android.util.LogUtil;
