@@ -2,11 +2,12 @@ package com.sola.v2ex_android.ui.base.view;
 
 import android.graphics.Rect;
 import android.os.Build;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.view.View;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 
 import com.sola.v2ex_android.ui.base.application.MyApplication;
 import com.sola.v2ex_android.util.DisplayUtil;
